+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_correct_edge_insets_constructor/prefer_correct_edge_insets_constructor_rule.dart
@@ -15,7 +15,7 @@ import '../../rule_utils.dart';
 part 'visitor.dart';
 
 class PreferCorrectEdgeInsetsConstructorRule extends FlutterRule {
-  static const ruleId = 'prefer-correct-edge-insets-constructor-rule';
+  static const ruleId = 'prefer-correct-edge-insets-constructor';
   static const _issueMessage = 'Prefer using correct EdgeInsets constructor.';
 
   PreferCorrectEdgeInsetsConstructorRule([