+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_correct_edge_insets_constructor/examples/example_ltrb.dart
@@ -1,0 +1,21 @@
+import 'flutter_defines.dart';
+
+class MyWidget extends StatelessWidget {
+  Widget build(BuildContext _) => Column(children: [
+        Container(
+          padding: const EdgeInsets.fromLTRB(1, 1, 0, 0), // LINT
+        ),
+        Container(
+          padding: const EdgeInsets.fromLTRB(12, 12, 12, 12), // LINT
+        ),
+        Container(
+          padding: const EdgeInsets.fromLTRB(1, 1, 1, 1), // LINT
+        ),
+        Container(
+          padding: const EdgeInsets.fromLTRB(3, 2, 3, 2), // LINT
+        ),
+        Container(
+          padding: const EdgeInsets.fromLTRB(3, 0, 0, 2), // LINT
+        ),
+      ]);
+}