+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ## 4.13.0
 
-* feat: add Checkstyle format reporter.
+* feat: add [Checkstyle](https://dartcodemetrics.dev/docs/cli/analyze#checkstyle) format reporter.
 * feat: add [prefer-immediate-return](https://dartcodemetrics.dev/docs/rules/common/prefer-immediate-return) rule
 
 ## 4.12.0