+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## 4.13.0
+## Unreleased
 
 * feat: add static code diagnostics `prefer-correct-edge-insets-constructor-rule`.
+
+## 4.13.0
+
 * feat: add [Checkstyle](https://dartcodemetrics.dev/docs/cli/analyze#checkstyle) format reporter.
 * feat: add [prefer-immediate-return](https://dartcodemetrics.dev/docs/rules/common/prefer-immediate-return) rule